pyproject.toml
[tool.poetry]
name = "esi-clone-status"
version = "0.1.1"
description = ""
authors = ["CoffeeAnon <dan@danjacobsen.com>"]

[tool.poetry.dependencies]
python = "^3.9"
gunicorn = "^20"
Flask-Login = "^0.6"
sqlalchemy = "^1.4"
simplejson = "^3.10"
Flask-Migrate = "^3.1.0"
Flask = "^2.1.2"
mypy = "^0.960"
python-dotenv = "^0.20.0"
EsiPy = "^1.2.3"
flask-swagger-ui = "^4.11.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
debugpy = "^1.6.3"
pytest = "^7.1.3"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"